pyproject.toml
[project]
name = "networksecurity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2025.8.3",
    "ipykernel>=6.30.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pymongo[srv]==3.11",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
]
